fix: redact proxy secrets in makeHttpRequest timeout errors too
Round 3 review: redactProxyUrlSecrets was only applied to the non-2xx
response error. makeHttpRequest's timeout error also embeds the raw URL
(which now carries the Scrapfly key + forwarded auth headers/body in its
query string) and flows into the queue retry/no-retries logs via
error.message.

Move redactProxyUrlSecrets into httpProxy.ts (re-exported from http.ts to
keep the http -> httpProxy import one-way, no cycle) and apply it to the
timeout error. All other URL log/trace paths use the bare this.url endpoint
(no secrets) after the earlier queryParams refactor.
