+++ src/httpProxy.ts
@@ -29,6 +29,36 @@ import {
  * @param options Request options
  * @returns Standard fetch Response
  */
+/**
+ * Redact secret query params from a proxy URL before it appears in logs or
+ * error messages. Proxy services (Scrapfly/CrawlBase) carry the API key — and,
+ * for Scrapfly, forwarded auth headers and request bodies — in the URL's query
+ * string, which would otherwise leak into error/retry logs on failure. Only the
+ * known proxy hosts and their sensitive params are touched; all other URLs are
+ * returned unchanged.
+ */
+export function redactProxyUrlSecrets(rawUrl: string): string {
+  try {
+    const u = new URL(rawUrl);
+    if (u.hostname === 'api.scrapfly.io') {
+      for (const name of [...u.searchParams.keys()]) {
+        const lower = name.toLowerCase();
+        if (lower === 'key' || lower === 'body' || lower.startsWith('headers[')) {
+          u.searchParams.set(name, '***');
+        }
+      }
+      return u.toString();
+    }
+    if (u.hostname === 'api.crawlbase.com' && u.searchParams.has('token')) {
+      u.searchParams.set('token', '***');
+      return u.toString();
+    }
+    return rawUrl;
+  } catch {
+    return rawUrl;
+  }
+}
+
 export async function makeHttpRequest(options: {
   method: string;
   url: string;
@@ -94,7 +124,7 @@ export async function makeHttpRequest(options: {
     // Surface timeouts with the same message shape we used before so callers
     // (and log greps) don't need to change.
     if (err?.name === 'TimeoutError' || err?.code === 'UND_ERR_ABORTED' || err?.name === 'AbortError') {
-      throw new Error(`HTTP request timed out after ${timeoutMs}ms: ${method} ${url}`);
+      throw new Error(`HTTP request timed out after ${timeoutMs}ms: ${method} ${redactProxyUrlSecrets(url)}`);
     }
     throw err;
   }
